# docs/openapi.yaml
...
openapi: "3.0.3"
info:
  title: Nuqta Backend API
  version: 1.0.0
  description: |
    REST JSON API for Nuqta POS / ERP system.
    All responses use the standard envelope:
    - Success: `{ ok: true, data: T }`
    - Error:   `{ ok: false, error: { code, message, details?, status? } }`
  contact:
    name: Nuqta Dev Team
servers:
  - url: /api/v1
    description: API v1 base path

# ─── Security Schemes ──────────────────────────────────────────────────
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT HS256 token.
        Access token TTL: 15 min (900s).
        Refresh token TTL: 7 days (604800s).
        Header: `Authorization: Bearer <accessToken>`

  # ─── Reusable Schemas ─────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - IDEMPOTENCY_REPLAY
            - INSUFFICIENT_STOCK
            - INVALID_STATE
            - INTERNAL_ERROR
        message:
          type: string
        details:
          description: Additional context (validation field errors, etc.)
        status:
          type: integer
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ApiError"

    # ── Pagination ──
...
    SaleItem:
      type: object
      properties:
        id: { type: integer }
        saleId: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitName: { type: string }
        unitFactor: { type: integer }
        quantityBase: { type: integer }
        batchId: { type: integer, nullable: true }
        unitPrice: { type: integer }
        discount: { type: integer }
        subtotal: { type: integer }
        cogs: { type: integer, nullable: true }
        weightedAverageCost: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
    Sale:
      type: object
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        customerId: { type: integer, nullable: true }
        subtotal: { type: integer }
        discount: { type: integer }
        tax: { type: integer }
        total: { type: integer }
        currency: { type: string }
        exchangeRate: { type: number }
        interestRate: { type: integer }
        interestAmount: { type: integer }
        paymentType: { type: string, enum: [cash, credit, mixed] }
        paidAmount: { type: integer }
        remainingAmount: { type: integer }
        status: { type: string, enum: [pending, completed, cancelled, refunded] }
        notes: { type: string, nullable: true }
        idempotencyKey: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
        createdBy: { type: integer, nullable: true }
        items: { type: array, items: { $ref: "#/components/schemas/SaleItem" } }
        cogs: { type: integer, nullable: true }
        totalCogs: { type: integer, nullable: true }
        profit: { type: integer, nullable: true }
        marginBps: { type: integer, nullable: true }
    CreateSaleItem:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId: { type: integer, minimum: 1 }
        quantity: { type: integer, minimum: 1 }
        unitPrice: { type: integer, minimum: 0 }
        discount: { type: integer, minimum: 0 }
        unitName: { type: string }
        unitFactor: { type: integer, minimum: 1 }
        batchId: { type: integer }
    CreateSaleRequest:
      type: object
      required: [items, paymentType]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/CreateSaleItem" }, minItems: 1 }
        customerId: { type: integer }
        discount: { type: integer, minimum: 0 }
        tax: { type: integer, minimum: 0 }
        paymentType: { type: string, enum: [cash, credit, mixed] }
        paidAmount: { type: integer, minimum: 0 }
        currency: { type: string }
        notes: { type: string }
        interestRate: { type: integer }
        interestRateBps: { type: integer }
        paymentMethod: { type: string, enum: [cash, card, bank_transfer, credit] }
        referenceNumber: { type: string }
        idempotencyKey: { type: string }
      additionalProperties: false
    AddPaymentRequest:
      type: object
      required: [amount, paymentMethod]
      properties:
        customerId: { type: integer }
        amount: { type: integer, minimum: 1 }
        currency: { type: string }
        exchangeRate: { type: number }
        paymentMethod: { type: string, enum: [cash, card, bank_transfer, credit] }
        referenceNumber: { type: string }
        notes: { type: string }
        idempotencyKey: { type: string }
      additionalProperties: false

    # ── Purchase ──
    Purchase:
      type: object
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        supplierId: { type: integer }
        subtotal: { type: integer }
        discount: { type: integer }
        tax: { type: integer }
        total: { type: integer }
        currency: { type: string }
        paidAmount: { type: integer }
        remainingAmount: { type: integer }
        status: { type: string, enum: [pending, completed, cancelled] }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
        createdBy: { type: integer, nullable: true }
        items: { type: array, items: {} }
    CreatePurchaseRequest:
      type: object
      required: [supplierId, items]
      properties:
        supplierId: { type: integer, minimum: 1 }
        items: { type: array, minItems: 1 }
        discount: { type: integer, minimum: 0 }
        tax: { type: integer, minimum: 0 }
        paidAmount: { type: integer, minimum: 0 }
        paymentMethod: { type: string, enum: [cash, card, bank_transfer, credit] }
        currency: { type: string }
        notes: { type: string }
        idempotencyKey: { type: string }
      additionalProperties: false

    # ── Accounting ──
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: "^\\d+$" }
      description: Numeric resource ID
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    OffsetParam:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    SearchParam:
      name: search
      in: query
      schema: { type: string }
...
    IdempotencyKeyHeader:
      name: X-Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Idempotency key — required for financial write operations

  # ─── Reusable Responses ───────────────────────────────────────────
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: Conflict (duplicate, state mismatch)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
paths:
...
  /sales:
    get:
      tags: [Sales]
      summary: List sales (paginated)
      operationId: listSales
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Paginated sales list
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Sale" } }
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Sales]
      summary: Create sale
      operationId: createSale
      description: |
        Creates a new sale with line items and payment information.
        Requires X-Idempotency-Key header to prevent double sales.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSaleRequest" }
      responses:
        "200":
          description: Sale created
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Sale" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /sales/{id}:
    get:
      tags: [Sales]
      summary: Get sale by ID
      operationId: getSale
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Sale details
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Sale" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /sales/{saleId}/payments:
    post:
      tags: [Sales]
      summary: Add payment to sale
      operationId: addSalePayment
      parameters:
        - name: saleId
          in: path
          required: true
          schema: { type: integer }
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddPaymentRequest" }
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /sales/{id}/cancel:
    post:
      tags: [Sales]
      summary: Cancel a sale
      operationId: cancelSale
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Sale cancelled
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Sale" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /sales/{id}/refund:
    post:
      tags: [Sales]
      summary: Refund a sale
      operationId: refundSale
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
                items: { type: array, items: { type: object, properties: { saleItemId: { type: integer }, quantity: { type: integer } } } }
      responses:
        "200":
          description: Sale refunded
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Sale" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /sales/{id}/receipt:
    get:
      tags: [Sales]
      summary: Get sale receipt data
      operationId: getSaleReceipt
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Receipt data
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data:
                    type: object
                    description: Full receipt with company info, sale details, items, payments
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── POS ──────────────────────────────────────────────────────────
...
  /purchases:
    get:
      tags: [Purchases]
      summary: List purchases (paginated)
      operationId: listPurchases
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Purchase list
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Purchase" } }
                      total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Purchases]
      summary: Create purchase
      operationId: createPurchase
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePurchaseRequest" }
      responses:
        "200":
          description: Purchase created
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Purchase" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /purchases/{id}:
    get:
      tags: [Purchases]
      summary: Get purchase by ID
      operationId: getPurchase
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Purchase details
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Purchase" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /purchases/{purchaseId}/payments:
    post:
      tags: [Purchases]
      summary: Add payment to purchase
      operationId: addPurchasePayment
      parameters:
        - name: purchaseId
          in: path
          required: true
          schema: { type: integer }
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddPaymentRequest" }
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ─── INVENTORY ────────────────────────────────────────────────────
...